♻️ refactor: 배포 상세 조회 파일 구조 정리 및 메시지 상수화
- 배포 상세 조회 관련 파일을 admin 디렉터리 구조로 정리 (views/services/serializers/tests)
- 스웨거 응답 description을 ErrorMessages 상수로 통일
- admin urls 라우팅 경로 업데이트
