+++ apps/exams/tests/admin/test_deployments_detail.py
